refactor: Carousel component to be in charge of managing cards of info openObserver to handle all init tasks
